:truck: (transit) Move energy, grid and simulation into their own file
